+++ __tests__/language_server/hover.test.ts
