+++ __tests__/language_server/prepare_rename.test.ts
