+++ __tests__/language_server/publish_diagnostics.test.ts
