+++ __tests__/language_server/semantic-token.test.ts
